+++ dsbridge/__init__.py
@@ -1,21 +1,25 @@
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
 from dsbridge.bridge import DSBridge
 from dsbridge.database import Base
 from dsbridge.database import Database
 from dsbridge.discord_bot import DiscordBot
 from dsbridge.models import ChatChannels
+from dsbridge.models import ChatChannelsMixin
 from dsbridge.models import Message
+from dsbridge.models import MessageMixin
 from dsbridge.models import utcnow
 from dsbridge.server import Server
 
 __all__ = [
     'Base',
     'ChatChannels',
+    'ChatChannelsMixin',
     'DSBridge',
     'Database',
     'DiscordBot',
     'Message',
+    'MessageMixin',
     'Server',
     'utcnow',
     '__version__',